requirements.txt
django
djangorestframework
django-allauth
django-rest-auth
djangorestframework-simplejwt
django-filter
django-cors-headers==4.3.1
corsheaders